requirements.txt
# AdTech Fraud Detection Platform - Core Dependencies
# ===================================================
# Install: pip install -r requirements.txt

# Snowflake
snowflake-snowpark-python>=1.15.0
snowflake-connector-python>=3.6.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0

# Logging
loguru>=0.7.0
